% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loop_cpp}
\alias{loop_cpp}
\title{A loop in c++ to speed up}
\usage{
loop_cpp(R, V)
}
\arguments{
\item{R}{a matrix (not sparse matrix, use \code{as.matrix(R)} to convert if necessary)}

\item{V}{a symmetric matrix (not sparse matrix)}
}
\description{
A loop in c++ to speed up
}
