% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partialR2}
\alias{partialR2}
\title{Partial R2}
\usage{
partialR2(mod, mod.r)
}
\arguments{
\item{mod}{a linear regression model}

\item{mod.r}{a reduced model based on \code{mod}}
}
\value{
R2 value between 0 and 1
}
\description{
Get partial R2 by comparing a model and its reduced model.
}
