% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partialR2adj}
\alias{partialR2adj}
\title{Adjusted partial R2}
\usage{
partialR2adj(mod, df.f, mod.r, df.r)
}
\arguments{
\item{mod}{a linear regression model}

\item{df.f}{degree of freedom of the \code{mod}}

\item{mod.r}{a reduced model based on \code{mod}}

\item{df.r}{degree of freedom of the reduced \code{mod.r}}
}
\value{
A list of both R2 and adjusted R2 , the latter is not necessary to be between 0 and 1.
}
\description{
Get adjusted partial R2 by comparing a model and its reduced model.
}
