% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phyloglm}
\alias{phyloglm}
\title{Phylogenetic GLM}
\usage{
phyloglm(formula, data = list(), phy, method = c("logistic_MPLE",
  "logistic_IG10", "poisson_GEE"), btol = 10, log.alpha.bound = 4,
  start.beta = NULL, start.alpha = NULL, boot = 0, full.matrix = TRUE,
  opt.method = c("Nelder-Mead", "L-BFGS-B", "BFGS", "SANN"))
}
\arguments{
\item{formula}{regression formula}

\item{data}{data frame to fit the model with}

\item{phy}{phylogenetic tree of type phylo with branch lengths}

\item{method}{The "logistic_IG10" method optimizes a GEE approximation to the penalized likelihood of the logistic regression. "logistic_MPLE" maximizes the penalized likelihood of the logistic regression. In both cases, the penalty is Firth's correction. The "poisson_GEE" method solves the generalized estimating equations (GEE) for Poisson regression.}

\item{btol}{(logistic regression only) bound on the linear predictor to bound the searching space.}

\item{log.alpha.bound}{(logistic regression only) bound for the log of the parameter alpha.}

\item{start.beta}{starting values for beta coefficients.}

\item{start.alpha}{(logistic regression only) starting values for alpha (phylogenetic correlation).}

\item{boot}{number of independent bootstrap replicates, \code{0} means no bootstrap.}

\item{full.matrix}{if \code{TRUE}, the full matrix of bootstrap estimates (coefficients and alpha) will be returned.}

\item{opt.method}{optim methods: "Nelder-Mead", "L-BFGS-B", "BFGS", "SANN"}
}
\value{
a large list with class as \code{phyloglm}
}
\description{
Fitting phylogenetic generalized linear models. Adapted from \code{phylolm::phyloglm()}.
}
\references{
See here for the orginal function \url{https://github.com/lamho86/phylolm/blob/master/man/phyloglm.Rd}
}
