% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_ce.R
\name{R2.ce}
\alias{R2.ce}
\title{Calculate R2.ce}
\usage{
R2.ce(mod = NULL, mod.r = NULL, phy = NULL, cpp = TRUE)
}
\arguments{
\item{mod}{a regression model with the following class: "lmerMod", "glmerMod", "phylolm", and "binaryPGLMM"}

\item{mod.r}{reduced model, if not provided, will use corresponding models with intercept as the only predictor}

\item{phy}{a phylogeny with tip labels and branch length}

\item{cpp}{whether to use Rcpp version of loops, default is TRUE}
}
\value{
R2.ce
}
\description{
Calculate R2.ce for LMM, GLMM, PGLM, and PGLMMs.
}
