% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_lr.R
\name{R2.lr}
\alias{R2.lr}
\title{Calculate R2.lr}
\usage{
R2.lr(mod = NULL, mod.r = NULL)
}
\arguments{
\item{mod}{a regression model with the following class: "lmerMod", "glmerMod", "phylolm", and "binaryPGLMM"}

\item{mod.r}{reduced model, if not provided, will use corresponding models with intercept as the only predictor}
}
\value{
R2.lr
}
\description{
Calculate R2.lr for LMM, GLMM, PGLM, and PGLMMs.
}
