% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_pred.R
\name{R2.pred}
\alias{R2.pred}
\title{Calculate R2.pred}
\usage{
R2.pred(mod = NULL, mod.r = NULL, phy = NULL, cpp = TRUE)
}
\arguments{
\item{mod}{a regression model with the following class: 'lmerMod', 'glmerMod', 'phylolm', and 'binaryPGLMM'}

\item{mod.r}{reduced model, if not provided, will use corresponding models with intercept as the only predictor}

\item{phy}{a phylogeny with tip labels and branch length}

\item{cpp}{whether to use Rcpp version of loops when the mod is class of 'phylolm', default is TRUE}
}
\value{
R2.pred
}
\description{
Calculate R2.pred for LMM, GLMM, PGLM, and PGLMMs.
}
