% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{R2}
\alias{R2}
\title{Calculate R2.lr, R2.ls, and R2.ce}
\usage{
R2(mod = NULL, mod.r = NULL, phy = NULL, lr = TRUE, ls = TRUE,
  ce = TRUE)
}
\arguments{
\item{mod}{a regression model with the following class: 'lmerMod', 'glmerMod', 'phylolm', and 'binaryPGLMM'}

\item{mod.r}{reduced model, if not provided, will use corresponding models with intercept as the only predictor}

\item{phy}{the phylogeny for phylogenetic models, which is not required to be specified for R2.lr.}

\item{lr}{whether to calculate R2.lr, default is TRUE}

\item{ls}{whether to calculate R2.ls, default is TRUE}

\item{ce}{whether to calculate R2.ce, default is TRUE}
}
\value{
all three R2s
}
\description{
Calculate all three R2s for LMM, GLMM, PGLM, and PGLMMs.
}
