% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\docType{package}
\name{rr2}
\alias{rr2}
\title{rr2: An R package to calculate R2s for regression models}
\description{
The rr2 package provides methods to calculate R2 for models with correlated errors, 
including Phylogenetic GLS, Phylogenetic Logistic Regression, LMMs, GLMM, and PGLMM.
}
