% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transf_phy}
\alias{transf_phy}
\title{Transform a phylogeny based on a phylolm model}
\usage{
transf_phy(phylolmMod, phy)
}
\arguments{
\item{phylolmMod}{A fitted phylolm model.}

\item{phy}{A phylogeny with class 'phylo'.}
}
\value{
A transformed phylogeny.
}
\description{
Using a fitted phylolm model to transform branch lengths of a phylogeny
}
