% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_lik.R
\name{R2.lik}
\alias{R2.lik}
\title{Calculate R2.lik}
\usage{
R2.lik(mod = NULL, mod.r = NULL)
}
\arguments{
\item{mod}{a regression model with the following class: 'lmerMod', 'glmerMod', 'phylolm', and 'phyloglm'}

\item{mod.r}{reduced model, if not provided, will use corresponding models with intercept as the only predictor}
}
\value{
R2.lik
}
\description{
Calculate R2.lik for LMM, GLMM, PGLM, and PGLMMs.
}
