% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_resid.R
\name{R2.resid}
\alias{R2.resid}
\title{Calculate R2.resid}
\usage{
R2.resid(mod = NULL, mod.r = NULL, phy = NULL)
}
\arguments{
\item{mod}{a regression model with the following class: 'lmerMod', 'glmerMod', 'phylolm', and 'binaryPGLMM'}

\item{mod.r}{reduced model, if not provided, will use corresponding models with intercept as the only predictor}

\item{phy}{a phylogeny with tip labels and branch length}
}
\value{
R2.resid
}
\description{
Calculate R2.resid for LMM, GLMM, PGLM, and PGLMMs.
}
