% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transf_phy}
\alias{transf_phy}
\title{Transform a phylogeny based on a phylolm model}
\usage{
transf_phy(phylolmMod, phy)
}
\arguments{
\item{phylolmMod}{a fitted phylolm model}

\item{phy}{a phylogeny with class "phylo"}
}
\value{
a transformed phylogeny
}
\description{
Using a fitted phylolm model to transform branch lengths of a phylogeny
}
